Reduce code to CKG requirements
This repo was duplicated from https://github.com/benjaminaaron/OntoEngine at commit 57e07f3 for the purpose of keeping the commit history. But with this commit a lot of things are being removed that are not necessary for the CKG to run. The CKG was just one experiment in the OntoEngine repo.